refactor: address paste-rule review feedback
- Downgrade the changeset bump for `@prosekit/core` and `prosekit` from
  `minor` to `patch`.
- Move `splitTextByRegex` (and its spec) into `packages/core/src/utils/`,
  and add test cases covering zero-length matches (`.?`, `a?`, `\d*`).
- Flatten the `extensions/paste-rule/` subdirectory so the paste-rule
  files live directly under `packages/core/src/extensions/`.
